fixup! Lazily create context attributes attached via baseplate.add_to_context
