--- baseplate/integration/mixins/__init__.py
@@ -1,5 +1,0 @@
-from .lazy_attributes import LazyAttributesMixin
-
-__all__ = [
-    'LazyAttributesMixin',
-]
